changed some stuff

+++ Calculator/MainWindow.xaml.cs
@@ -91,7 +91,7 @@ namespace Calculator
             {
                 if (!input.Text.Equals("0"))
                 {
-                    if ((input.Text.Length == 2 && Decimal.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) < 0) || input.Text.Length == 1)
+                    if ((input.Text.Length == 2 && Double.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) < 0) || input.Text.Length == 1)
                     {
                         input.Text = "0";
                     }
@@ -213,9 +213,9 @@ namespace Calculator
         private void Negation_Click(object sender, RoutedEventArgs e)
         {
             Error();
-            if (Decimal.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) != 0)
+            if (Double.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) != 0)
             {
-                if (Decimal.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) < 0)
+                if (Double.Parse(input.Text.Replace(',', '.'), CultureInfo.InvariantCulture) < 0)
                 {
                     input.Text = input.Text.Remove(0, 1);
                 }
